chore: 도메인 별 폴더링 적용
