+++ src/main/java/gift/product/controller/auth/AdminAuthController.java
@@ -1,26 +1,18 @@
-package gift.product.controller;
+package gift.product.controller.auth;
 
-import gift.product.dto.AdminProductDto;
-import gift.product.dto.JwtResponse;
-import gift.product.dto.MemberDto;
-import gift.product.model.Product;
+import gift.product.dto.auth.JwtResponse;
+import gift.product.dto.auth.MemberDto;
 import gift.product.service.AuthService;
-import gift.product.service.ProductService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import java.net.URI;
-import java.util.List;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -28,15 +20,12 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminController {
+public class AdminAuthController {
 
     private final String REDIRECT_ADMIN_LOGIN_PROCESSING = "redirect:/admin/login/process";
-    private final String REDIRECT_ADMIN_PRODUCTS = "redirect:/admin/products";
-    private final ProductService productService;
     private final AuthService authService;
 
-    public AdminController(ProductService productService, AuthService authService) {
-        this.productService = productService;
+    public AdminAuthController(AuthService authService) {
         this.authService = authService;
     }
 
@@ -69,45 +58,6 @@ public class AdminController {
         HttpHeaders headers = new HttpHeaders();
         headers.setLocation(URI.create("/admin/products"));
         return new ResponseEntity<>(headers, HttpStatus.FOUND);
-    }
-
-
-    @GetMapping("/products")
-    public String products(Model model) {
-        List<Product> products = productService.getProductAll();
-        model.addAttribute("products", products);
-        return "admin/products";
-    }
-
-    @GetMapping("/products/insert")
-    public String insertForm() {
-        return "admin/insertForm";
-    }
-
-    @PostMapping("/products/insert")
-    public String insertProduct(@Valid AdminProductDto adminProductDto) {
-        productService.insertProduct(adminProductDto);
-        return REDIRECT_ADMIN_PRODUCTS;
-    }
-
-    @GetMapping("/products/update/{id}")
-    public String updateForm(@PathVariable(name = "id") Long productId, Model model) {
-        Product product = productService.getProduct(productId);
-        model.addAttribute("product", product);
-        return "admin/updateForm";
-    }
-
-    @PutMapping("/products/update/{id}")
-    public String updateProduct(@PathVariable(name = "id") Long productId,
-        @Valid AdminProductDto adminProductDto) {
-        productService.updateProduct(productId, adminProductDto);
-        return REDIRECT_ADMIN_PRODUCTS;
-    }
-
-    @DeleteMapping("/products/delete/{id}")
-    public String deleteProduct(@PathVariable(name = "id") Long productId) {
-        productService.deleteProduct(productId);
-        return REDIRECT_ADMIN_PRODUCTS;
     }
 
     private void addAccessTokenCookieInResponse(JwtResponse jwtResponse,
