+++ src/main/java/gift/product/controller/category/CategoryController.java
@@ -1,7 +1,7 @@
-package gift.product.controller;
+package gift.product.controller.category;
 
 
-import gift.product.dto.CategoryDto;
+import gift.product.dto.category.CategoryDto;
 import gift.product.model.Category;
 import gift.product.service.CategoryService;
 import java.util.List;
