+++ src/main/java/gift/product/dto/auth/AccessAndRefreshToken.java
@@ -1,4 +1,4 @@
-package gift.product.dto;
+package gift.product.dto.auth;
 
 public record AccessAndRefreshToken(
     String accessToken,
