+++ src/main/java/gift/product/dto/category/CategoryDto.java
@@ -1,4 +1,4 @@
-package gift.product.dto;
+package gift.product.dto.category;
 
 public record CategoryDto(
     String name
