+++ src/main/java/gift/product/dto/option/OptionResponse.java
@@ -1,4 +1,4 @@
-package gift.product.dto;
+package gift.product.dto.option;
 
 public record OptionResponse(
     Long id,
