+++ src/main/java/gift/product/dto/option/OptionSubtractAmount.java
@@ -1,4 +1,4 @@
-package gift.product.dto;
+package gift.product.dto.option;
 
 public record OptionSubtractAmount(
     int amount
