+++ src/main/java/gift/product/dto/wish/WishDto.java
@@ -1,4 +1,4 @@
-package gift.product.dto;
+package gift.product.dto.wish;
 
 public record WishDto(
     Long productId
