+++ src/main/java/gift/product/service/WishService.java
@@ -1,7 +1,7 @@
 package gift.product.service;
 
-import gift.product.dto.LoginMember;
-import gift.product.dto.WishDto;
+import gift.product.dto.auth.LoginMember;
+import gift.product.dto.wish.WishDto;
 import gift.product.model.Member;
 import gift.product.model.Product;
 import gift.product.model.Wish;
