+++ src/test/java/gift/service/OptionServiceTest.java
@@ -6,9 +6,9 @@ import static org.mockito.BDDMockito.clearInvocations;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 
-import gift.product.dto.OptionDto;
-import gift.product.dto.OptionResponse;
-import gift.product.dto.OptionSubtractAmount;
+import gift.product.dto.option.OptionDto;
+import gift.product.dto.option.OptionResponse;
+import gift.product.dto.option.OptionSubtractAmount;
 import gift.product.model.Category;
 import gift.product.model.Option;
 import gift.product.model.Product;
